requirements.txt
Flask
FeedGenerator
requests
